Clearly separate different parts of the exports macro
Serializing and emitting the generated metadata is now no longer part of
collecting the metadata and generating the scaffolding code, respectively.
